requirements.txt
autopep8==2.0.1
pre-commit==3.0.4
ipykernel==6.21.1
transformers==4.42.4
torch==2.2.2
pillow==10.4.0